updated car profile controller
