+++ profile-service/src/main/resources/db/migration/V1.1__init.sql
@@ -1,13 +1,13 @@
 CREATE TABLE `crs_profile_schema`.`CRS_CUSTOMER` (
-  `CUSTOMER_ID` INT NOT NULL,
+  `CUSTOMER_ID` BIGINT NOT NULL,
   `CUSTOMER_NAME` VARCHAR(45) NULL,
-  `CUSTOMER_AGE` INT NULL,
-  `CUSTOMER_GENDER` VARCHAR(45) NULL,
-  `CUSTOMER_PHONE` VARCHAR(45) NULL,
+  `CUSTOMER_AGE` TINYINT(4) NULL,
+  `CUSTOMER_GENDER` ENUM('MALE', 'FEMALE') NULL,
+  `CUSTOMER_PHONE` VARCHAR(10) NULL,
   `CUSTOMER_ADDRESS` VARCHAR(145) NULL,
-  `CUSTOMER_LIECENCE` VARCHAR(45) NULL,
-  `CUSTOMER_PAN` VARCHAR(45) NULL,
-  `CUSTOMER_ADHAAR` VARCHAR(45) NULL,
+  `CUSTOMER_LIECENCE` VARCHAR(25) NULL,
+  `CUSTOMER_PAN` VARCHAR(10) NULL,
+  `CUSTOMER_ADHAAR` VARCHAR(16) NULL,
   `CUSTOMER_REG_DATE` DATETIME NULL,
   PRIMARY KEY (`CUSTOMER_ID`));
 
@@ -15,14 +15,14 @@ ALTER TABLE `crs_profile_schema`.`CRS_CUSTOMER`
 ADD COLUMN `CUSTOMER_EMAIL` VARCHAR(45) NULL AFTER `CUSTOMER_PHONE`;
 
 CREATE TABLE `crs_profile_schema`.`CRS_VENDOR` (
-  `VENDOR_ID` INT NOT NULL,
+  `VENDOR_ID` BIGINT NOT NULL,
   `VENDOR_NAME` VARCHAR(45) NULL,
-  `VENDOR_AGE` VARCHAR(45) NULL,
-  `VENDOR_GENDER` VARCHAR(45) NULL,
+  `VENDOR_AGE` TINYINT(4) NULL,
+  `VENDOR_GENDER` ENUM('MALE', 'FEMALE') NULL,
   `VENDOR_PHONE` VARCHAR(45) NULL,
   `VENDOR_ADDRESS` VARCHAR(140) NULL,
-  `VENDOR_PAN` VARCHAR(45) NULL,
-  `VENDOR_ADHAAR` VARCHAR(45) NULL,
+  `VENDOR_PAN` VARCHAR(10) NULL,
+  `VENDOR_ADHAAR` VARCHAR(16) NULL,
   `VENDOR_REG_DATE` DATETIME NULL,
   `VENDOR_GSTIN` VARCHAR(45) NULL,
   PRIMARY KEY (`VENDOR_ID`));
@@ -64,7 +64,7 @@ CREATE TABLE `crs_profile_schema`.`CRS_CARS` (
   `CAR_ACTIVE` BIT NULL,
   `CAR_PRICE_PER_DAY` INT NULL,
   `CAR_COLOR` VARCHAR(45) NULL,
-  `CAR_VENDOR_ID` INT NULL,
+  `CAR_VENDOR_ID` BIGINT NULL,
   `CAR_LOCATION_ID` INT NULL,
   `CAR_REG_ON` DATETIME NULL,
   PRIMARY KEY (`CAR_ID`),
